+++ app/assets/stylesheets/components/_index.scss
@@ -7,3 +7,4 @@
 @import "compass";
 @import "loadbar";
 @import "plants";
+@import "overlays";
