+++ app/assets/stylesheets/components/_overlays.scss
@@ -1,0 +1,38 @@
+.overlay {
+  height: 0;
+  width: 100%;
+  background-color: white;
+  position: fixed;
+  z-index: 2;
+  bottom: 0;
+  left: 0;
+  border-radius: 1rem 1rem 0 0;
+  transition: height 0.2s ease-in-out;
+  padding: 1rem;
+  &.active {
+    height: 80vh;
+  }
+}
+
+.overlay-filter {
+  position: fixed;
+  z-index: -1;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  transition: filter 0.3s ease-in-out;
+  backdrop-filter: blur(5px);
+  &.active {
+    z-index: 1;
+  }
+}
+
+.doc-inside{
+  position: relative;
+  display: flex;
+  text-align: left;
+  flex-direction: column;
+  justify-content: center;
+  align-items: left;
+}
